Update @ckeditor/ckeditor-core to v27
