+++ types/ckeditor__ckeditor5-core/OTHER_FILES.txt
@@ -1,0 +1,2 @@
+src/editor/editorwithui.d.ts
+src/pendingactions.d.ts
