+++ types/ckeditor__ckeditor5-core/src/command.d.ts
@@ -1,0 +1,51 @@
+import { Emitter, EmitterMixinDelegateChain } from "@ckeditor/ckeditor5-utils/src/emittermixin";
+import EventInfo from "@ckeditor/ckeditor5-utils/src/eventinfo";
+import { BindChain, Observable } from "@ckeditor/ckeditor5-utils/src/observablemixin";
+import Editor from "./editor/editor";
+import * as engine from "@ckeditor/ckeditor5-engine";
+import { PriorityString } from "@ckeditor/ckeditor5-utils/src/priorities";
+
+export default class Command implements Emitter, Observable {
+    readonly value?: unknown;
+    readonly editor: Editor;
+    readonly isEnabled: boolean;
+
+    constructor(editor: Editor);
+    clearForceDisabled(id: string): void;
+    destroy(): void;
+    execute(options?: Record<string, unknown>): void;
+    forceDisabled(id: string): void;
+    refresh(): void;
+
+    set(option: Record<string, unknown>): void;
+    set(name: string, value: unknown): void;
+    bind(...bindProperties: string[]): BindChain;
+    unbind(...unbindProperties: string[]): void;
+    decorate(methodName: string): void;
+
+    on: (
+        event: string,
+        callback: (info: EventInfo<Emitter>, data: engine.view.observer.DomEventData) => void,
+        options?: { priority: PriorityString | number },
+    ) => void;
+    once(
+        event: string,
+        callback: (info: EventInfo, data: engine.view.observer.DomEventData) => void,
+        options?: { priority: PriorityString | number },
+    ): void;
+    off(event: string, callback?: (info: EventInfo, data: engine.view.observer.DomEventData) => void): void;
+    listenTo(
+        emitter: Emitter,
+        event: string,
+        callback: (info: EventInfo, data: engine.view.observer.DomEventData) => void,
+        options?: { priority?: PriorityString | number },
+    ): void;
+    stopListening(
+        emitter?: Emitter,
+        event?: string,
+        callback?: (info: EventInfo, data: engine.view.observer.DomEventData) => void,
+    ): void;
+    fire(eventOrInfo: string | EventInfo<Emitter>, ...args: any[]): any;
+    delegate(...events: string[]): EmitterMixinDelegateChain;
+    stopDelegating(event?: string, emitter?: Emitter): void;
+}
