+++ types/ckeditor__ckeditor5-core/src/editingkeystrokehandler.d.ts
@@ -1,0 +1,8 @@
+import KeystrokeHandler from "@ckeditor/ckeditor5-utils/src/keystrokehandler";
+import Editor from "./editor/editor";
+
+export default class EditingKeystrokeHandler extends KeystrokeHandler {
+    readonly editor: Editor;
+
+    constructor(editor: Editor);
+}
