+++ types/ckeditor__ckeditor5-core/src/editor/editorwithui.d.ts
@@ -1,0 +1,5 @@
+import EditorUI from "./editorui";
+
+export interface EditorWithUI {
+    readonly ui: EditorUI;
+}
