+++ types/ckeditor__ckeditor5-core/src/editor/utils/dataapimixin.d.ts
@@ -1,0 +1,8 @@
+export interface DataApi {
+    setData(data: string): void;
+    getData(options?: { rootName?: string; trim?: "empty" | "none" }): string;
+}
+
+declare const DataApiMixin: DataApi;
+
+export default DataApiMixin;
