+++ types/ckeditor__ckeditor5-core/src/editor/utils/elementapimixin.d.ts
@@ -1,0 +1,8 @@
+export interface ElementApi {
+    readonly sourceElement?: HTMLElement;
+    updateSourceElement(): void;
+}
+
+declare const ElementApiMixin: ElementApi;
+
+export default ElementApiMixin;
