+++ types/ckeditor__ckeditor5-core/src/editor/utils/securesourceelement.d.ts
@@ -1,0 +1,3 @@
+import Editor from "../editor";
+
+export default function secureSourceElement(editor: Editor): void;
