+++ types/ckeditor__ckeditor5-core/v11/tsconfig.json
@@ -11,14 +11,20 @@
         "noImplicitThis": true,
         "strictNullChecks": true,
         "strictFunctionTypes": true,
-        "baseUrl": "../",
+        "baseUrl": "../../",
         "typeRoots": [
-            "../"
+            "../../"
         ],
         "types": [],
         "noEmit": true,
         "forceConsistentCasingInFileNames": true,
         "paths": {
+            "@ckeditor/ckeditor5-core": [
+                "ckeditor__ckeditor5-core/v11"
+            ],
+            "@ckeditor/ckeditor5-utils": [
+                "ckeditor__ckeditor5-utils/v10"
+            ],
             "@ckeditor/*": [
                 "ckeditor__*"
             ]
